adjusting file structure
