+++ predictionguard/src/completions.py
@@ -4,7 +4,7 @@ import requests
 from typing import Any, Dict, List, Optional, Union
 from warnings import warn
 
-from ...version import __version__
+from ..version import __version__
 
 
 class Completions:
